data_consumer/requirements.txt
kafka-python
pyspark
pymongo
pandas